* 'master' of git://scm.dev.nokia.troll.no/qt/qt-qml-team: (766 commits)
  Avoid spurious detaching in QDir::to/fromNativeSeparators
  Use fromLocal8Bit for reversing toLocal8Bit
  Don't convert template's path separators again
  Changed if/if/else/if/eleven chain to switch
  Modulus of negative dividends is undefined or negative
  Fix warning when compiling with VS 2008
  Re-introduce Q_GLOBAL_STATIC_INIT
  Don't leak MouseEvents
  Fix race condition in processJobs()
  HTTP backend: invoke methods directly for synchronous requests
  Don't endless loop
  uikit: update SDK setting for the simulator
  Add option to compare to different platform in lancelot autotest
  uikit: Use release Qt for simulator, update README
  Automatic closing of input panel on uikit when QML item looses focus.
  Compile on symbian
  Don't release register prematurely
  Fixed compile error in qwindowsurface_qws.cpp.
  Fixed regression introduced by 5842d19cf3dff37a85c.
  Remove DerivedSources.pro from WebKit.
  ...

+++ src/corelib/tools/qbytearray.cpp
@@ -541,7 +541,7 @@ QByteArray qUncompress(const uchar* data, int nbytes)
 
     forever {
         ulong alloc = len;
-        if (len  >= (1 << 31) - sizeof(QByteArray::Data)) {
+        if (len  >= ulong(1 << 31) - sizeof(QByteArray::Data)) {
             //QByteArray does not support that huge size anyway.
             qWarning("qUncompress: Input data is corrupted");
             return QByteArray();
@@ -561,7 +561,7 @@ QByteArray qUncompress(const uchar* data, int nbytes)
         switch (res) {
         case Z_OK:
             if (len != alloc) {
-                if (len  >= (1 << 31) - sizeof(QByteArray::Data)) {
+                if (len  >= ulong(1 << 31) - sizeof(QByteArray::Data)) {
                     //QByteArray does not support that huge size anyway.
                     qWarning("qUncompress: Input data is corrupted");
                     return QByteArray();
